requirements.txt
notebook
# pip install torch==1.11.0+cu113 --extra-index-url https://download.pytorch.org/whl/cu113
torch
# pip install -Iv torch-scatter torch-sparse==0.6.13 torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.11.0+cu113.html
pypots
pytorch-lightning
#pytorch-forecasting

pandas

hydra-core
hydra-colorlog
hydra-optuna-sweeper

tensorboard
pyrootutils
rich
seaborn
#tods
#pyod
